requirements.txt
paho-mqtt
pytest
schedule
